tests/wpt/web-platform-tests/content-security-policy/style-src/style-src-3_1.html

<!doctype html>
<html>
<head>
    <title></title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>

    <style>
        /* none of this should be applied */
        #log {
            margin-left: 200px;
        }
    </style>
</head>
<body>
    <h1>
        Inline style should not be applied
        without unsafe-inline directive
    </h1>
    <div id='log'></div>

    <script>
    test(function() {
        var logEl = document.getElementById("log");
        var marginLeftVal = getComputedStyle(logEl).getPropertyValue('margin-left');
        assert_false(marginLeftVal == "200px")},
        "Inline style should not be applied"
    );
    </script>

    <script async defer src='../support/checkReport.sub.js?reportField=violated-directive&reportValue=style-src%20%27self%27'></script>
</body>
</html>
